requirements.txt
alembic==1.13.1
asttokens==2.4.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docker==7.0.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.9.0
Flask==3.0.2
fonttools==4.49.0
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lightgbm==4.3.0
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mlflow==2.10.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
querystring-parser==1.2.4
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.27
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.3.0
torch==2.2.0+cpu
torchaudio==2.2.0+cpu
torchvision==0.17.0+cpu
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
ucimlrepo==0.0.3
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
xgboost==2.0.3
zipp==3.17.0
